+++ python_modules/jupedsim/README.md
@@ -1,3 +1,5 @@
 ## Testing
 
 from ```jpscore/python_modules/jupedsim``` you can call ```python -m pytest```
+
+to test the Distribution visually call ```python -m streamlit``` from ```jpscore/python_modules/jupedsim```
